Update wdw.sh
In this updated version of the script, the process_login function takes an extra argument that indicates whether the user is currently logged in. If this argument is 1, the function calculates the duration of the user's session and includes it in the output. 

The date command is used to convert the login time to a Unix timestamp, and the difference between this timestamp and the current time is calculated to determine the duration of the session. 

The duration is then formatted as hours, minutes, and seconds for display.